+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 ## [Unreleased]
 
 ### Added
+- Added support for sound alerts on timers (Thanks @muyiwaolurin!)
 
 ### Changed
 
 ### Fixed
+- Fixed customize menu to display actual workflow names
 
 ## [0.5.0] - 2025-12-03